refactor: move trpc from /server to /server/api
