+++ src/app/api/trpc/[trpc]/route.ts
@@ -1,8 +1,8 @@
 import { fetchRequestHandler } from '@trpc/server/adapters/fetch';
 import { NextRequest } from 'next/server';
 
-import { appRouter } from '@/server';
-import { createTRPCContext } from '@/server/trpc';
+import { appRouter } from '@/server/api';
+import { createTRPCContext } from '@/server/api/trpc';
 
 const createContext = async (req: NextRequest) => {
   return createTRPCContext({
